requirements.txt
torch>=2.1
transformers>=4.51
datasets
accelerate
bitsandbytes
rouge-score
nltk
scikit-learn
tensorboard
wandb
pyyaml
numpy
tqdm
